Feat: Implement abstract State class
States are stored in a polymorphic, intrusive linked list. Is this the
right data structure? Well I personally prefer it to
std::vector<std::unique_ptr<State>> as it avoids [re]allocations, and
means the list can be modified while it's traversed, but whether it's
worth the additional effort to construct this manually is up for a lot
of debate. It's fun though!