fix: Correctly nul-terminate text passed to espeak-ng
This fixes an issue where espeak would overread the buffer and speak
undefined characters from memory.

Speaker::speak() now has to take a const std::string&, as it needs to be
nul-terminated. I could've used const char * but that would've made the
callsites really ugly. An overload could always be added if necessary.